pyproject.toml
[project]
name = "helm-charts-docs"
version = "0.1.0"
description = "Documentation for Helm Charts repository"
readme = "docs/index.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocs-minify-plugin>=0.8.0",
    "pymdown-extensions>=10.0",
    "pygments>=2.17.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "copier>=9.0.0",
    "pre-commit>=3.0.0",
    "ruff>=0.15.7",
]

# ─────────────────────────────────────────────────────────────────────────────
# Pytest Configuration
# ─────────────────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests (may interact with system)",
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_file = ".logs/pytest/test.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s [%(levelname)8s] %(name)s (%(filename)s:%(lineno)d): %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"

# ─────────────────────────────────────────────────────────────────────────────
# Ruff Configuration
# ─────────────────────────────────────────────────────────────────────────────

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
per-file-ignores = {"scripts/sync-chart-docs.py" = ["E402"], "scripts/generate-docs.py" = ["E501"]}
